support to load multiple target definition files at runtime
This update allows to add additional target definitions at runtime. The
limit for the summ of all targets within all loaded files is 100.

In the exampleET only the chip target is active. Once double tapped on
the screen an additional target file is loaded and the other two
targets become functional.